moving path-details to impl-file
